Refactor: 시그니처 변경
- #8
