+++ src/test/java/co/wanted/board/api/auth/application/encode/PasswordValidatorTest.java
@@ -1,4 +1,4 @@
-package co.wanted.board.api.auth.encode.application;
+package co.wanted.board.api.auth.application.encode;
 
 import co.wanted.board.api.member.domain.Member;
 import co.wanted.board.api.member.domain.Password;
@@ -7,10 +7,12 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest
+@ActiveProfiles("test")
 class PasswordValidatorTest {
 
     Member testMember;
